Moves module
